+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## Last Changes
 
+- [#4](https://github.com/x1B/nbe/issues/4): Allow to remove selected edge using DELETE key.
 - [#1](https://github.com/x1B/nbe/issues/1): Replaced embedded dependencies with bower.